+++ package.json
@@ -6,9 +6,9 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start-server": "ts-node index.ts",
-    "migration-run": "typeorm-ts-node-esm migration:run -d datasource.ts",
-    "migration-generate": "typeorm-ts-node-esm migration:generate -d datasource.ts",
-    "start-client": "ts-node client/client.ts"
+    "migration-run": "typeorm-ts-node-esm migration:run -d src/datasource.ts",
+    "migration-generate": "typeorm-ts-node-esm migration:generate -d src/datasource.ts",
+    "start-client": "ts-node src/client/client.ts"
   },
   "author": "",
   "license": "ISC",
