+++ src/datasource.ts
@@ -8,7 +8,7 @@ export const AppDataSource = new DataSource({
     username: process.env.PG_USER,
     password: process.env.PG_PASSWORD,
     database: process.env.PG_DATABASE,
-    entities: ['src/entities/*.{js,ts}'],
+    entities: ['entities/*.{js,ts}'],
     synchronize: false,
-    migrations: ['src/migration/*.{js,ts}'],
+    migrations: ['migration/*.{js,ts}'],
 });
