+++ tsconfig.json
@@ -4,14 +4,11 @@
     "lib": ["ES6", "dom"],                                        /* Specify a set of bundled library declaration files that describe the target runtime environment. */
     "experimentalDecorators": true,                   /* Enable experimental support for legacy experimental decorators. */
     "emitDecoratorMetadata": true,                    /* Emit design-type metadata for decorated declarations in source files. */
-
     "module": "commonjs",                                /* Specify what module code is generated. */
     "moduleResolution": "node",                     /* Specify how TypeScript looks up a file from a given module specifier. */
     "esModuleInterop": true,                             /* Emit additional JavaScript to ease support for importing CommonJS modules. This enables 'allowSyntheticDefaultImports' for type compatibility. */
     "forceConsistentCasingInFileNames": true,            /* Ensure that casing is correct in imports. */
-
     "strict": true,                                      /* Enable all strict type-checking options. */
-   
     "skipLibCheck": true                                 /* Skip type checking all .d.ts files. */
   }
 }
